refeactor the code and make a markdown previwer components .
